requirements.txt
attrs
requests
bs4